fix(adminrole): 将 AdminRole ID 类型改为 int64 并完善数据初始化
- 将 AdminRole 相关所有 ID 字段和函数参数从 int 改为 int64，以支持更大范围的 ID 值
- 在数据库迁移代码中同步更改列类型及注释，保持一致性
- 添加 AdminRole.ID 字段的验证器，并在创建更新中进行校验
- 改进种子数据初始化流程，新增初始化超级管理员用户和角色的逻辑
- 完善默认权限的初始化，包括菜单和按钮权限的创建及关联
- 在迁移命令中添加初始化数据失败的错误处理与日志输出

+++ ent/adminrole.go
@@ -16,7 +16,8 @@ import (
 type AdminRole struct {
 	config `json:"-"`
 	// ID of the ent.
-	ID int `json:"id,omitempty"`
+	// 主键ID，自增
+	ID int64 `json:"id,omitempty"`
 	// 创建时间
 	CreatedAt time.Time `json:"created_at,omitempty"`
 	// 更新时间
@@ -73,7 +74,7 @@ func (_m *AdminRole) assignValues(columns []string, values []any) error {
 			if !ok {
 				return fmt.Errorf("unexpected type %T for field id", value)
 			}
-			_m.ID = int(value.Int64)
+			_m.ID = int64(value.Int64)
 		case adminrole.FieldCreatedAt:
 			if value, ok := values[i].(*sql.NullTime); !ok {
 				return fmt.Errorf("unexpected type %T for field created_at", values[i])
